+++ tests/functions/test_complete.py
@@ -23,12 +23,12 @@ def df1():
 
 
 def test_empty_column(df1):
-    "Return dataframe if `columns` is empty."
+    """Return dataframe if `columns` is empty."""
     assert_frame_equal(df1.complete(), df1)
 
 
 def test_MultiIndex_column(df1):
-    "Raise ValueError if column is a MultiIndex."
+    """Raise ValueError if column is a MultiIndex."""
     df = df1
     df.columns = [["A", "B", "C"], list(df.columns)]
     with pytest.raises(ValueError):
@@ -36,7 +36,7 @@ def test_MultiIndex_column(df1):
 
 
 def test_column_duplicated(df1):
-    "Raise ValueError if column is duplicated in `columns`"
+    """Raise ValueError if column is duplicated in `columns`"""
     with pytest.raises(ValueError):
         df1.complete(
             columns=[
@@ -48,21 +48,30 @@ def test_column_duplicated(df1):
 
 
 def test_type_columns(df1):
-    "Raise error if columns is not a list object."
+    """Raise error if columns is not a list object."""
     with pytest.raises(TypeError):
         df1.complete(columns="Year")
 
 
 def test_fill_value_is_a_dict(df1):
-    "Raise error if fill_value is not a dictionary"
+    """Raise error if fill_value is not a dictionary"""
     with pytest.raises(TypeError):
         df1.complete(columns=["Year", "Taxon"], fill_value=0)
 
 
 def test_wrong_column_fill_value(df1):
-    "Raise ValueError if column in `fill_value` does not exist."
+    """Raise ValueError if column in `fill_value` does not exist."""
     with pytest.raises(ValueError):
         df1.complete(columns=["Taxon", "Year"], fill_value={"year": 0})
+
+
+def test_wrong_data_type_dict(df1):
+    """
+    Raise ValueError if value in dictionary
+    is not a 1-dimensional object.
+    """
+    with pytest.raises(ValueError):
+        df1.complete(columns=[{"Year": pd.DataFrame([2005, 2006, 2007])}])
 
 
 frame = pd.DataFrame(
@@ -222,6 +231,39 @@ def test_dict_series_duplicates(df1, df1_output):
     assert_frame_equal(result, df1_output)
 
 
+def test_dict_values_outside_range(df1):
+    """
+    Test the output if a dictionary is present,
+    and none of the values in the dataframe,
+    for the corresponding label, is not present
+    in the dictionary's values.
+    """
+    result = df1.complete(
+        columns=[("Taxon", "Abundance"), {"Year": np.arange(2005, 2007)}]
+    )
+    expected = pd.DataFrame(
+        [
+            {"Taxon": "Agarum", "Abundance": 1, "Year": 1999},
+            {"Taxon": "Agarum", "Abundance": 1, "Year": 2005},
+            {"Taxon": "Agarum", "Abundance": 1, "Year": 2006},
+            {"Taxon": "Agarum", "Abundance": 8, "Year": 2004},
+            {"Taxon": "Agarum", "Abundance": 8, "Year": 2005},
+            {"Taxon": "Agarum", "Abundance": 8, "Year": 2006},
+            {"Taxon": "Saccharina", "Abundance": 2, "Year": 2004},
+            {"Taxon": "Saccharina", "Abundance": 2, "Year": 2005},
+            {"Taxon": "Saccharina", "Abundance": 2, "Year": 2006},
+            {"Taxon": "Saccharina", "Abundance": 4, "Year": 1999},
+            {"Taxon": "Saccharina", "Abundance": 4, "Year": 2005},
+            {"Taxon": "Saccharina", "Abundance": 4, "Year": 2006},
+            {"Taxon": "Saccharina", "Abundance": 5, "Year": 2000},
+            {"Taxon": "Saccharina", "Abundance": 5, "Year": 2005},
+            {"Taxon": "Saccharina", "Abundance": 5, "Year": 2006},
+        ]
+    )
+
+    assert_frame_equal(result, expected)
+
+
 # adapted from https://tidyr.tidyverse.org/reference/complete.html
 complete_parameters = [
     (
@@ -293,28 +335,9 @@ def test_complete(df, columns, output):
     assert_frame_equal(df.complete(columns), output)
 
 
-# https://stackoverflow.com/questions/63541729/
-# pandas-how-to-include-all-columns-for-all-rows-although-value-is-missing-in-a-d
-# /63543164#63543164
-def test_duplicate_index():
-    """Test that the complete function works for duplicate index."""
-    df = pd.DataFrame(
-        {
-            "row": {
-                0: "21.08.2020",
-                1: "21.08.2020",
-                2: "21.08.2020",
-                3: "21.08.2020",
-                4: "22.08.2020",
-                5: "22.08.2020",
-                6: "22.08.2020",
-            },
-            "column": {0: "A", 1: "A", 2: "B", 3: "C", 4: "A", 5: "B", 6: "B"},
-            "value": {0: 43, 1: 36, 2: 36, 3: 28, 4: 16, 5: 40, 6: 34},
-        }
-    )
-
-    dup_expected_output = pd.DataFrame(
+@pytest.fixture
+def duplicates():
+    return pd.DataFrame(
         {
             "row": [
                 "21.08.2020",
@@ -331,9 +354,63 @@ def test_duplicate_index():
         }
     )
 
+
+# https://stackoverflow.com/questions/63541729/
+# pandas-how-to-include-all-columns-for-all-rows-although-value-is-missing-in-a-d
+# /63543164#63543164
+def test_duplicates(duplicates):
+    """Test that the complete function works for duplicate values."""
+    df = pd.DataFrame(
+        {
+            "row": {
+                0: "21.08.2020",
+                1: "21.08.2020",
+                2: "21.08.2020",
+                3: "21.08.2020",
+                4: "22.08.2020",
+                5: "22.08.2020",
+                6: "22.08.2020",
+            },
+            "column": {0: "A", 1: "A", 2: "B", 3: "C", 4: "A", 5: "B", 6: "B"},
+            "value": {0: 43, 1: 36, 2: 36, 3: 28, 4: 16, 5: 40, 6: 34},
+        }
+    )
+
     result = df.complete(columns=["row", "column"], fill_value={"value": 0})
 
-    assert_frame_equal(result, dup_expected_output)
+    assert_frame_equal(result, duplicates)
+
+
+def test_unsorted_duplicates(duplicates):
+    """Test output for unsorted duplicates."""
+
+    df = pd.DataFrame(
+        {
+            "row": {
+                0: "22.08.2020",
+                1: "22.08.2020",
+                2: "21.08.2020",
+                3: "21.08.2020",
+                4: "21.08.2020",
+                5: "21.08.2020",
+                6: "22.08.2020",
+            },
+            "column": {
+                0: "B",
+                1: "B",
+                2: "A",
+                3: "A",
+                4: "B",
+                5: "C",
+                6: "A",
+            },
+            "value": {0: 40, 1: 34, 2: 43, 3: 36, 4: 36, 5: 28, 6: 16},
+        }
+    )
+
+    result = df.complete(columns=["row", "column"], fill_value={"value": 0})
+
+    assert_frame_equal(result, duplicates)
 
 
 # https://stackoverflow.com/questions/32874239/
@@ -359,7 +436,6 @@ def test_grouping_first_columns():
     output2 = pd.DataFrame(
         {
             "id": [1, 1, 1, 2, 2, 2, 3, 3, 3],
-            "choice": [5, 6, 7, 5, 6, 7, 5, 6, 7],
             "c": [9.0, 9.0, 9.0, np.nan, np.nan, np.nan, 11.0, 11.0, 11.0],
             "d": [
                 pd.NaT,
@@ -372,6 +448,7 @@ def test_grouping_first_columns():
                 pd.Timestamp("2015-09-29 00:00:00"),
                 pd.Timestamp("2015-09-29 00:00:00"),
             ],
+            "choice": [5, 6, 7, 5, 6, 7, 5, 6, 7],
         }
     )
     result = df2.complete(columns=[("id", "c", "d"), "choice"])
@@ -393,17 +470,109 @@ def test_complete_multiple_groupings():
 
     output3 = pd.DataFrame(
         {
-            "project_id": [1, 1, 1, 1, 1, 2, 2, 2, 2, 2],
-            "meta": ["A", "A", "B", "B", "C", "A", "A", "B", "B", "C"],
-            "domain1": ["d", "e", "h", "i", "k", "d", "e", "h", "i", "k"],
-            "question_count": [3, 3, 3, 3, 3, 2, 2, 2, 2, 2],
-            "tag_count": [2.0, 1.0, 3.0, 2.0, 0.0, 1.0, 0.0, 0.0, 1.0, 2.0],
+            "meta": ["A", "A", "A", "A", "B", "B", "B", "B", "C", "C"],
+            "domain1": ["d", "d", "e", "e", "h", "h", "i", "i", "k", "k"],
+            "project_id": [1, 2, 1, 2, 1, 2, 1, 2, 1, 2],
+            "question_count": [3, 2, 3, 2, 3, 2, 3, 2, 3, 2],
+            "tag_count": [2.0, 1.0, 1.0, 0.0, 3.0, 0.0, 2.0, 1.0, 0.0, 2.0],
         }
     )
 
     result = df3.complete(
         columns=[("meta", "domain1"), ("project_id", "question_count")],
         fill_value={"tag_count": 0},
-    ).sort_values("project_id", ignore_index=True)
-
+    )
     assert_frame_equal(result, output3)
+
+
+@pytest.fixture
+def output_dict_tuple():
+    return pd.DataFrame(
+        [
+            {"Year": 1999, "Taxon": "Agarum", "Abundance": 1},
+            {"Year": 1999, "Taxon": "Agarum", "Abundance": 8},
+            {"Year": 1999, "Taxon": "Saccharina", "Abundance": 2},
+            {"Year": 1999, "Taxon": "Saccharina", "Abundance": 4},
+            {"Year": 1999, "Taxon": "Saccharina", "Abundance": 5},
+            {"Year": 2000, "Taxon": "Agarum", "Abundance": 1},
+            {"Year": 2000, "Taxon": "Agarum", "Abundance": 8},
+            {"Year": 2000, "Taxon": "Saccharina", "Abundance": 2},
+            {"Year": 2000, "Taxon": "Saccharina", "Abundance": 4},
+            {"Year": 2000, "Taxon": "Saccharina", "Abundance": 5},
+            {"Year": 2001, "Taxon": "Agarum", "Abundance": 1},
+            {"Year": 2001, "Taxon": "Agarum", "Abundance": 8},
+            {"Year": 2001, "Taxon": "Saccharina", "Abundance": 2},
+            {"Year": 2001, "Taxon": "Saccharina", "Abundance": 4},
+            {"Year": 2001, "Taxon": "Saccharina", "Abundance": 5},
+            {"Year": 2002, "Taxon": "Agarum", "Abundance": 1},
+            {"Year": 2002, "Taxon": "Agarum", "Abundance": 8},
+            {"Year": 2002, "Taxon": "Saccharina", "Abundance": 2},
+            {"Year": 2002, "Taxon": "Saccharina", "Abundance": 4},
+            {"Year": 2002, "Taxon": "Saccharina", "Abundance": 5},
+            {"Year": 2003, "Taxon": "Agarum", "Abundance": 1},
+            {"Year": 2003, "Taxon": "Agarum", "Abundance": 8},
+            {"Year": 2003, "Taxon": "Saccharina", "Abundance": 2},
+            {"Year": 2003, "Taxon": "Saccharina", "Abundance": 4},
+            {"Year": 2003, "Taxon": "Saccharina", "Abundance": 5},
+            {"Year": 2004, "Taxon": "Agarum", "Abundance": 1},
+            {"Year": 2004, "Taxon": "Agarum", "Abundance": 8},
+            {"Year": 2004, "Taxon": "Saccharina", "Abundance": 2},
+            {"Year": 2004, "Taxon": "Saccharina", "Abundance": 4},
+            {"Year": 2004, "Taxon": "Saccharina", "Abundance": 5},
+        ]
+    )
+
+
+def test_dict_tuple(df1, output_dict_tuple):
+    """
+    Test if a dictionary and a tuple/list
+    are included in the `columns` parameter.
+    """
+
+    result = df1.complete(
+        columns=[
+            {"Year": lambda x: range(x.Year.min(), x.Year.max() + 1)},
+            ("Taxon", "Abundance"),
+        ]
+    )
+
+    assert_frame_equal(result, output_dict_tuple)
+
+
+def test_complete_groupby():
+    """Test output in the presence of a groupby."""
+    df = pd.DataFrame(
+        {
+            "state": ["CA", "CA", "HI", "HI", "HI", "NY", "NY"],
+            "year": [2010, 2013, 2010, 2012, 2016, 2009, 2013],
+            "value": [1, 3, 1, 2, 3, 2, 5],
+        }
+    )
+
+    result = df.complete(
+        columns=[{"year": lambda x: range(x.year.min(), x.year.max() + 1)}],
+        by="state",
+    )
+
+    expected = pd.DataFrame(
+        [
+            {"state": "CA", "year": 2010, "value": 1.0},
+            {"state": "CA", "year": 2011, "value": np.nan},
+            {"state": "CA", "year": 2012, "value": np.nan},
+            {"state": "CA", "year": 2013, "value": 3.0},
+            {"state": "HI", "year": 2010, "value": 1.0},
+            {"state": "HI", "year": 2011, "value": np.nan},
+            {"state": "HI", "year": 2012, "value": 2.0},
+            {"state": "HI", "year": 2013, "value": np.nan},
+            {"state": "HI", "year": 2014, "value": np.nan},
+            {"state": "HI", "year": 2015, "value": np.nan},
+            {"state": "HI", "year": 2016, "value": 3.0},
+            {"state": "NY", "year": 2009, "value": 2.0},
+            {"state": "NY", "year": 2010, "value": np.nan},
+            {"state": "NY", "year": 2011, "value": np.nan},
+            {"state": "NY", "year": 2012, "value": np.nan},
+            {"state": "NY", "year": 2013, "value": 5.0},
+        ]
+    )
+
+    assert_frame_equal(result, expected)